Renamed testing utils
