+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 testpaths = tests
 python_files = test_*.py
-norecursedirs = utils
 markers =
     first: mark a test as first to be run.
+addopts = --ignore=tests/utils/
