+++ tests/test_KoalaBot.py
@@ -18,8 +18,8 @@ from discord.ext import commands
 
 # Own modules
 import KoalaBot
-from tests.utils.TestUtils import FakeAuthor
-from tests.utils.TestUtilsCog import TestUtilsCog
+from tests.utils_testing.TestUtils import FakeAuthor
+from tests.utils_testing.LastCtxCog import LastCtxCog
 from utils.KoalaDBManager import KoalaDBManager
 
 # Constants
@@ -33,7 +33,7 @@ DBManager.create_base_tables()
 @pytest.fixture(autouse=True)
 async def test_ctx(bot):
     global utils_cog
-    utils_cog = TestUtilsCog(bot)
+    utils_cog = LastCtxCog(bot)
     bot.add_cog(utils_cog)
     dpytest.configure(bot)
     await dpytest.message(KoalaBot.COMMAND_PREFIX + "store_ctx")
